added css for heading and move the songs folder into components
